+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Version 0.11.4
 
 * Improve AST node generation compatibility for Python 3.14+
+* Load `spec/spec_helper.py` and `specs/spec_helper.py` by path (doesn't require `spec` to be a Python package)
 * Add Python 3.13 and 3.14 trove classifiers
 
 ## Version 0.11.3